requirements.txt
pillow~=9.0.1
reportlab
requests~=2.25.1
pandas~=2.2.3
gspread~=6.2.0
oauth2client~=4.1.3
beautifulsoup4~=4.10.0
argparse~=1.4.0
fpdf2~=2.8.2